#  ddtrace/tracer: switch to vmihailenco/msgpack encoding library