… ddtrace/vendor/github.com/ugorji/go/.travis.yml
… ddtrace/vendor/github.com/ugorji/go/LICENSE
… ddtrace/vendor/github.com/ugorji/go/README.md
… ddtrace/vendor/github.com/ugorji/go/codec/0doc.go
… ddtrace/vendor/github.com/ugorji/go/codec/README.md
… ddtrace/vendor/github.com/ugorji/go/codec/binc.go
… ddtrace/vendor/github.com/ugorji/go/codec/cbor.go
… ddtrace/vendor/github.com/ugorji/go/codec/cbor_test.go
… ddtrace/vendor/github.com/ugorji/go/codec/codec_test.go
… ddtrace/vendor/github.com/ugorji/go/codec/codecgen/README.md
… ddtrace/vendor/github.com/ugorji/go/codec/codecgen/gen.go
… ddtrace/vendor/github.com/ugorji/go/codec/codecgen/z.go
… ddtrace/vendor/github.com/ugorji/go/codec/decode.go
… ddtrace/vendor/github.com/ugorji/go/codec/encode.go
… ddtrace/vendor/github.com/ugorji/go/codec/fast-path.generated.go
… ddtrace/vendor/github.com/ugorji/go/codec/fast-path.go.tmpl
… ddtrace/vendor/github.com/ugorji/go/codec/fast-path.not.go
… ddtrace/vendor/github.com/ugorji/go/codec/gen-dec-array.go.tmpl
… ddtrace/vendor/github.com/ugorji/go/codec/gen-dec-map.go.tmpl
… ddtrace/vendor/github.com/ugorji/go/codec/gen-enc-chan.go.tmpl
… ddtrace/vendor/github.com/ugorji/go/codec/gen-helper.generated.go
… ddtrace/vendor/github.com/ugorji/go/codec/gen-helper.go.tmpl
… ddtrace/vendor/github.com/ugorji/go/codec/gen.generated.go
… ddtrace/vendor/github.com/ugorji/go/codec/gen.go
… ddtrace/vendor/github.com/ugorji/go/codec/goversion_arrayof_gte_go15.go
… ddtrace/vendor/github.com/ugorji/go/codec/goversion_arrayof_lt_go15.go
… ddtrace/vendor/github.com/ugorji/go/codec/goversion_makemap_gte_go19.go
… ddtrace/vendor/github.com/ugorji/go/codec/goversion_makemap_lt_go19.go
… ddtrace/vendor/github.com/ugorji/go/codec/goversion_unexportedembeddedptr_gte_go110.go
… ddtrace/vendor/github.com/ugorji/go/codec/goversion_unexportedembeddedptr_lt_go110.go