```
name                      old time/op    new time/op    delta
PayloadThroughput/10K-4     19.5µs ± 0%    15.4µs ± 0%     -21.05%
PayloadThroughput/100K-4    88.7µs ± 0%   138.0µs ± 0%     +55.56%
PayloadThroughput/1MB-4     2.41ms ± 0%    1.00ms ± 0%     -58.29%

name                      old speed      new speed      delta
PayloadThroughput/10K-4    531MB/s ± 0%   673MB/s ± 0%     +26.74%
PayloadThroughput/100K-4  1.17GB/s ± 0%  0.75GB/s ± 0%     -35.68%
PayloadThroughput/1MB-4    430MB/s ± 0%  1032MB/s ± 0%    +139.91%

name                      old alloc/op   new alloc/op   delta
PayloadThroughput/10K-4     35.6kB ± 0%    27.3kB ± 0%     -23.26%
PayloadThroughput/100K-4     232kB ± 0%     364kB ± 0%     +56.99%
PayloadThroughput/1MB-4     2.85MB ± 0%    3.64MB ± 0%     +27.82%

name                      old allocs/op  new allocs/op  delta
PayloadThroughput/10K-4       2.00 ± 0%      4.00 ± 0%    +100.00%
PayloadThroughput/100K-4      2.00 ± 0%     31.00 ± 0%   +1450.00%
PayloadThroughput/1MB-4       2.00 ± 0%    301.00 ± 0%  +14950.00%
```